Moved id3 examples to id3_algorithm class.
